// src/enums/pagesEnus.ts
export enum Pages {
  Gender = 'gender',
  Relationship = 'relationship',
  Parent = 'parent',
  SingleParent = 'single-parent',
  RelationshipProblem = 'relationship-problem',
  SingleProblem = 'single-problem',
  Overthink = 'overthink',
  IntrovetExtrovert = 'introvet-extrovert',
  PartnerGender = 'partner-gender',
  ImportantTraits = 'important-traits',
  EmotionalControl = 'emotional-control',
  PartnerPriority = 'partner-priority',
  RelationshipGoals = 'relationship-goals',
  AboutUs = 'about-us',
  HowItWorks = 'how-it-works',
  Result = 'result',
  Fancy = 'fancy',
}